Initial implementation of the new logcat panel.
This logcat panel will show up in ddms if it is launched with -Dcom.android.ddms.useNewLogCatView.
Currently, this implementation only displays a table with a list of logcat messages. It lacks
features such as filters, go-to source of exception, export to log, user preferences, etc.
Future patches should enhance the UI to achieve feature parity with older UI.

Change-Id: I3dde3c590c839318ce57bb5f005627f580ebb06c