Format newly created files
This changeset hooks the XML formatter up to the New XML File and New
Android Project wizards such that the newly created XML files are run
through the formatter, thereby picking up whatever formatting
customizations the user has made.

It also makes it such that when you finish the New XML File wizard for
a file that does not have a graphical editor, the caret position is
shown and the view receives focus so you can start editing at the
appropriate position.

Finally, it tweaks the defaut root element for color-lists.

Change-Id: I19606a085d6df132009679ddfb4c91152cbdc453