SdkMan2: Fix edge case when install/delete packages.
One issue when installing or deleting packages is that
sometimes after an install the display would show both
the uninstalled and installed package. And when deleting
it would also show both states.
The core issue is that it's a bad idea to rely on
Set<>.contains() when your objects equality isn't quite
what the hash code says.

Or said otherwise, there's more than one definition
of "equals()" for Package objects. That is when a
package is installed or deleted, its internal state
changes and its archive changes (it looses or gain an
OS path), which changes the overall hash code but
at that point it's the same object from the user point
of view.

Anyway, long story made short: this fixes that.

Change-Id: I6ec705cc807d96d097cf9ea7faf8234f7ed34416