Improve partial XML formatting
This changeset improves the behavior of the formatting when formatting
just a part of the document (e.g. the selected region, or more
accurately the nodes overlapping the selected region.)

First, it looks up the current indentation levels in the document of
the start node and its parent and uses that for the indentation, which
means that nodes appear to be pretty printed in line instead of
jumping to whatever the absolute indentation level would correspond
to.

Second, the formatter now has a mode where it can format only the
opening tag of an element. This is a special case of formatting a node
which is particularly useful when you enable automatic formatting by
the layout editor, since just setting a property on a root element
shouldn't reformat all the element children of the edited node.

Third, there are various tweaks to how the whitespace is handled
around the inserted node. It also fixes a couple of bugs.

Change-Id: If1dd2dc5091a08b770927cd4b9fcafe0c032a216