Fix template creation of ids
We have a quickfix for creating missing resources, such as
@string/foo.  This code was not working correctly for @id resources,
because ids use a different declaration syntax (there is no value to
be specified as a text node child of the item tag, and the type
attribute should be specified).

This changeset also fixes ResourceHelper's isFileBasedResourceType().
ResourceHelper.isFileBasedResourceType recently started returning true
for ResourceType.ID. This seems related to a change in the return
values from FolderTypeRelationship.

In any case, ids should not be treated as #isFileBasedResourceType,
because that method is used to determine whether the id is one-to-one
with a filename, which in turn means that the resource name validator
will insist that the filename can only contain -lowercase- characters
(plus digits and underscores).

Incidentally this change was caught by a unit test, yay!

Change-Id: I0c95c20d666d0744fda73e8230846d24ddd437c3