Reformat the ADT plugin.xml file
The plugin.xml file had a lot of very inconsistent formatting,
primarily because the Eclipse Format action seems to be disabled in
it, and most of the contents are the result of copy pasting in
snippets to register various services based on code samples in
documentation.

This changeset simply reformats the file (4 spaces, one attribute per
line) to make the format consistent and easier to adhere to when
inserting new content.

There are NO semantic changes - it's a pure reformat only affecting
whitespace.

Change-Id: I42237f67ada0e2ddeffc5e305a7df1083f0fdb97