Update XML Mac keybindings for select enclosing and select last
One of the most useful selection operations in Eclipse Java editors is
"select enclosing" and "select last".  The keybindings on Mac are
Ctrl-Shift-Up and Ctrl-Shift-Down. However, in XML files, those
keybindings do something different (making the caret jump, without
selection, to the previous element instead of the parent).

It turns out the XML editor also implements those actions, but bound
to other keys on the Mac. The reason for this is that the Java editor
applies special mac specific keybindings for those actions, and the
XML editor does not.

This changeset updates the keybindings, for Mac only, to the same
keybindings as the Mac-only Java editor bindings for the same actions.

Change-Id: I81f7f6d14fe8dd4a79aeace757d6c25c011eedb8