Ability to suppress rendering fidelity warnings
If you're working with a custom component that uses unsupported
graphics operations, the layout editor will show an error area. This
error is always there. That can be annoying since there's nothing you
can do about it after reporting a bug.

This changeset adds the ability to suppress these errors for the
current IDE session. It adds a "Ignore" hyperlink next to each
fidelity rendering error message, and when you click it, the specific
type of rendering error is suppressed and hidden for the current IDE
session.

These warnings are only suppressed for a single session and not
permanently. This is because you are suppressing an error not for a
specific layout but for any occurrence of that error. If these ignore
settings were persisted you might open some unrelated layout in the
future and be puzzled why it does not look right.

Change-Id: I2d20a36aee7767257b16ad805b0e8da36a6eb97e