SdkMan2: Correctly handle local items newer than remote.
This fixes an edge case where a local package is newer
(e.g. has a higher revision number) than a remote one.
This typically happens during development and QA but
not in a normal user scenario.
In this case the newer local package 'hides' the remote
package so that we don't see a dup in the list.

Also fixes the source filter when merging remote packages
and add a comment on the rationale of which, namely that
if two remote repos were to publish the same packages we
DO want to see dups. This is not supposed to happen and
we would want to see it and have it fixed.

Change-Id: Iab51cfd64b4078f3233872973173465a924eac1e