Fix logcat table scrolling behavior.
This patch makes the logcat table scroll behavior to be consistent
with other consoles/terminals: The table will scroll to reveal a
newly added message if and only if the last logcat message is
currently displayed (which is the same as saying if the scrollbar
is at the bottom of the table). If the user scrolls up the table
by moving the scroll thumb, then newly added messages will not
trigger a scroll.

Change-Id: Ia9c180b093de91406e8ced020589c71b4d2e1eb1