Add-ons do not support codenames.
There was a typo in the repository XSD and the addon XSD
that had an optional 'codename' element to the addon type.
Since it's optional and (should have) never beend used,
it's safe to remove it and document why we're removing it
without spinning a new XSD version.
Also added a unit-test to enforce it's invalid to have.

Change-Id: Id3c5725c022d86ac0f46532efe25ac5c000c1af7