Add support for live filter/search bar.
A search bar at the top of the logcat view provides a quick way to
search for messages matching a certain pattern. By default, words in
the query are searched in the logcat message text. Other fields such
as pid or tag can be searched by using a prefix such as pid: or tag:.

Change-Id: I3f9f83697f84c30b0371beedfba457480001d71a