Pull View Rules API into a separate library
This changeset pulls out the API classes from the plugin sources and
into a separate standalone .jar library. The library depends on the
common.jar library.

With the separate view API it should be possible to build designtime
helpers (view rules) for custom views to improve editing behavior in
the layout editor.

Change-Id: I20bb511668de2fe52910e5fe0bbd3ec2a18b5a08