ADT: Fix manifest editor marking files as dirty on open.
Originally this was supposed to be about fixing
issue 18689 (can't edit/save minSdkVersion) but I
can't repro that anymore, so instead I focused on
fixing the marked-dirty not going away. That one is
due to the manifest descriptor lacking the XMLNS
for "android". Upon loading a custom hidden attribute
was created and marked as dirty, which is wrong.

Change-Id: I9f0d3809632fdbeef8c8a39f79c62914818b1efe