Use Eclipse XML parser instead of JDK parser when formatting
This changeset adds a utility method which parses an XML string using
the Eclipse structured model instead of the JDK parser. The resulting
document has some extra metadata which the XML formatter can use. In
particular, it distinguishes between empty tags and tags without
children.

Now that this is fixed the formatting on project creation code
includes manifest files, which I excluded previously because it made
the <uses-sdk> element look ugly.

Change-Id: I7069fa43df33d874303435b3625f1c7b20f5b3d3