Add Unwrap Refactoring
This changeset adds a new visual refactoring: "Remove Container". This
is basically the reverse of the "Wrap In" refactoring (which surrounds
the selected views with a new parent layout). Instead it takes a
parent and removes it, in the process transferring the namespace
declaration (if a root element) and any layout params to the children.

Change-Id: Id535b88fcf39a8b7a83fd6feb981fff000140e8a