Precompilation only executes AAPT when necessary
This change adds resource tracking to the ResourceManager.
Each ResourceRepository now has new methods:

void markForIdRefresh() to set the repository as "dirty"
boolean needsIdRefresh() to check whether the repository is dirty
void setIdsRefreshed() to set the repository as "clean"

During the precompilation step, the PreCompiler will query the
ResourceManager to see if any of the repositories included in the build
are marked as dirty. AAPT will only be run if one or more dirty repositories
are found.

Repositories are marked as clean when R.java is regenerated and IDs are
set in ProjectResources.

Change-Id: I575ab819702508eacd247b282c3de8979f2f0ab9