ADT: Fix issue 18689 (can't edit/save minSdkVersion)
This actually fixes the issue directly, not by a side effect.
For some weird reason, I was always creating UI nodes for
unknown attributes with a setDirty(true). That's obviously
wrong, the attribute should not be set dirty when created
but only when a value is assigned to it, if needed.

Change-Id: I4d4b6dbbc0ab7e837a772994b3e6bf6c2236de65