Fix issue 19057: Drag & drop within canvas broken on Linux
This changeset fixes
  http://code.google.com/p/android/issues/detail?id=19057

by working around the fact that on some Linux/SWT versions, mouse up
events are delivered during a drag and drop. These events would
interfere with the normal drag & drop machinery. Luckily we can tell
whether or not we're in the middle of a drag & drop operation, and if
so ignore the mouse up event completely.

Change-Id: I183e0b5e360865724b87cc200e1899a2b31cc2ef