Misc refactoring and test fixes
This changeset turns on automatic formatting of the affected regions
in extract include, change layout, wrap in and style refactorings.

It also fixes a couple of minor bugs - one handling empty tags in the
change widget type refactoring, and one handling selection offsets.

It also turns on automatic formatting of GUI editor changes (there was
a previous changeset which prepared for this but forgot to actually
switch the default option.)

This changeset also makes various test infrastructure fixes: code
completion tests recently stopped working for a number of files
(something about the initialization order of editor types when running
as tests).  I also tweaked the goldenfile format of the diffs to avoid
trailing diffs on lines with just a ">" or "<" since gerrit makes thee
light up in red.

Change-Id: I78b858f0b21e5b2f25a9566ad495bad09c66a18f