Modify android create/update project to deal with new build.xml
New mechanism to embed a build version into build.xml to figure out
whether the build.xml file is obsolete or not.
Use the version-tag: custom case to prevent erase user modifications.

Fixed some issues with the name of the project (put in build.xml) for
test projects which don't have any activity.

Change-Id: Ib0333d15ca9558fffe24a718e704fcf4176c9014