Add ia32-libs hint on Linux when aapt fails to run
I had difficulties getting ADT to work when I tried running it on a 64
bit Linux system.  Aapt couldn't run, even though I verified that it
was there. After searching a bit, I not only found a bunch of forums
talking about this error (and offering pretty complex solutions), it
turns out this is even discussed in the SDK Installation
Troubleshooting guide. The solution is simple: Install 32-bit
libraries using

   sudo apt-get install ia32-libs

This changeset looks for this specific scenario in the aapt error
handler and emits a second error line which suggests that this
solution might be needed.

Change-Id: I453f99df61a198352f2cea4fe5cb7c517eeaf8c8