Workaround for Linux painting issue
This fixes:
http://code.google.com/p/android/issues/detail?id=19447

Turns out that on Linux, painting an image with a scale transform can
be extremely expensive - it can take multiple seconds, which makes for
example interactive dragging in a RelativeLayout over a zoomed canvas
appear not to work at all (because in some cases each drag update
takes up to ten seconds).

We rely on scaling image painting when the canvas is zoomed.

To work around this, we "prescale" the images instead: Using the
source AWT image, apply a scale once (whenever the image or the zoom
level has changed), and then store the scaled image as a plain SWT
image which we can then paint on each paint refresh using a normal
(not scaling) image paint.

This change also revealed that during a palette drag, the source image
would get trashed by the palette preview code because the editor XML
model was marked as modified. This is not necessary, so during preview
painting the editor is instructed to ignore XML model updates.

Change-Id: I78f3adb77754d4f9b0daaeeb662d53f1d7ed6042