Move creation of LogCatReceiver to a separate factory.
Until now there was only one instance of LogCatReceiver. This
instance would listen to the logcat output from the currently
selected device.

This patch modifies this behavior to make one LogCatReceiver
responsible for listening to only one device. If there are multiple
devices, then multiple instances of LogCatReceivers are instantiated.
The instantiation is also moved to a separate factory.

This modification allows for the scenario where the LogCatReceiver's
can be instantiated independent of the UI, and are reused by the UI
whenever needed.

Change-Id: I46008cbd5ee41880372a62e7e09634ce2d2fc122