SDK Manager: Fix log printf calls.
The SDK manager used to have log(String) calls which were
replaced by log(printfFormat, printfArgs..) calls in R12.

In a few places, we pass a string to the printf format that
comes from some user variable or URL or another method and
this string might contain % occurences that should not be
taken as formatters.

Change-Id: Ida29a18ba70d869a4f9d66f1d66edc010d9ec62d