Change default launcher icon name in new Android projects
When creating a new project, the default project template includes a
launcher icon named "icon.png". This changeset changes the name of
this default icon to "ic_launcher.png", in accordance with the
recommended naming convention listed in

http://developer.android.com/guide/practices/ui_guidelines/icon_design.html

(The secondary motivation for this is that the Icon Set Wizard will
suggest ic_launcher as a default launcher name, and it would be nice
if going through the wizard with the default settings would replace
the current project's launcher icon. Given the guidelines above I'd
rather not change the wizard default name to "icon")

Change-Id: I9f516c769019c2635edad4e1e6f01a5110d156fe