Use project minimum SDK to generate fewer icons
This changeset makes the icon generator consider the project's minimum
SDK version, to only create necessary icons.  For example, by default
the tab icon generator will generate different icons for v5 and newer
and folder older icons. With this CL, if the minimum SDK is 5 or
higher, then only the newer style icons are created (and they are
placed in the default drawable folder instead of a -v5 version of the
folder). Similarly, notification icons for v11, v9 and older are also
created conditionally.

Change-Id: Id96d572d4b8f23cd59d0f1b983c6aea85683eb0f