Copy OpenGLES emulation libraries to SDK installation.
This patch ensures that the following host shared libraries
are copies to $SDK/tools/lib, when building the SDK:

  - libOpenglRender
  - libEGL_translator
  - libGLES_CM_translator
  - libGLES_V2_translator

Note that the Windows DLL must be copied manually in patch_windows_sdk.sh
I didn't find a way to make it work using tools.atree.

This has been tested to build the Linux, Windows and OS X SDK images.

Change-Id: Idae8fe705befbbcc31e8d4ff71998ea0803b32f1