Reduce the number of UI refreshes of the logcat table.
Currently, every time a logcat message is received, the UI is refreshed.
If multiple messages are received in close succession, multiple refreshes
are enqueued to the UI thread to be run. This results in the UI spending
all its time refreshing the table.

This patch adds a refresh to the queue only if there are no pending
refreshes. We can do so since any refresh automatically picks up the latest
state of the logcat messages received.

Change-Id: I2d9df679299710e54986c86a3610d3546924aba8