Use project selector button in the New XML File wizard
I changed the project chooser in the New Android Icon Wizard from a
text field to a button (which shows the project name and icon, and
when clicked brings up the project selector dialog). This helps draw
the focus down to the next text field (the name) so you don't
accidentally try to enter it in the project field, and with the icon
it's also more obvious that it's the place where the project appears.

This CL makes the same change to the New XML File wizard.

Change-Id: Ib6496cb5b803f36228eb94fb5b1b2d0441f82ff8