SDK Manager: allow manual install of incompatible archives.
(where 'incompatible' means not intended for the host OS,
e.g. installing an archive flagged as Linux on Windows.)

By default, SDK Manager 2 doesn't show archives and
focuses on packages. However some vendors might want to
create e.g. an add-on with an OS-specific archive. Users
that do not have the OS need a way to still access and
install this. The way to do is to enable 'Show Archive
Details' in the Packages menu, manually select the
archive for any OS and install it.

SDK Bug: 19725

Change-Id: I619c9b882a4c56c0dc67b0073e6e3386f6a0e215