SDK Manager2: fix large downloads.
UrlOpener was using a BufferedHtmlEntity reader.
As a consequence its getContent() was loading large packages
in memory and resulting in OutOfMemory exceptions, whilst at
the same time blocking and thus rendering the download
progress report useless. This fixes it by simply passing
through in the underlying stream and freeing resources when
the stream is closed.

Change-Id: Ib3b265953931fecdf202197230ee9f24d8178faa