Make the dependency graph use extension restrictions per folder.
Before there was a way to filter touched file by extension(s) but
this was for all input folders.

Now each input folder can have a different set of extension restrictions.
This will allow apkbuilder to use DependencyGraph and extension restriction
for all its input folders.

Also fixed an issue where aapt would not package the resources if an
xml files was touched. This is because it didn't make a difference
between compiling the resources for IDs and packaging the resource
values.

Change-Id: I797d3a24c6c1f999e9d412c4ff8aa826ed16fc09