Eliminate GEF dependency
ADT used to depend on GEF (the Graphical Editing Framework in
Eclipse), but with GLE2 it no longer does. However, there were still a
couple of references left to GEF (a javadoc reference, some usages of
its Rectangle class, and a reference in the manifest).

Not all Eclipse distributions ship with GEF, which means some users
get a missing dependency error which is not immediately obvious to
those not familiar with Eclipse libraries.

Since we don't depend on GEF for anything important, eliminate the
last dependencies.

Change-Id: I5d1e11a195da496dc10c10fc7865df7e4d183629