Only mark downstream projects for build when they are open
When a resource is modified in a library project, then all downstream
projects which depends on this library project are also marked with a
persistent property which indicates that they need to run aapt as part
of their next build.

However, if a parent project is closed, this would throw an exception
since you can only set properties on open projects.

Change-Id: I20f88bb28dde43b4f8bba427f45449245badbdb5