Misc GridLayout interaction fixes
A couple of fixes to the interaction support for GridLayouts.

If you resize within a GridLayout and drag only a horizontal or
vertical edge, then make the resize message only list the current new
size in the resized dimension. In other words, if you drag the right
hand side edge, it might say "wrap_content", but if you drag the
bottom right corner, it might say "50dp x wrap_content". This is
similar to what is done in for example the LinearLayout during
resizing.

Second, prefer matching with an existing row or column over adding a
new row/column adjacent to a right edge.

Third, fix drag feedback messages to be more accurate when dealing
with adjacent matches.

Change-Id: I2e09731712c961dfd8b9c60fb33d37f0bc0a1cc4