Add/Remove wrap_content when converting GridLayouts
In GridLayout, the layout_width and layout_height attributes use
wrap_content by default. We already avoid setting these on the
elements themselves when dragging into the layout, but this changeset
makes the Change Layout refactoring remove wrap_content attributes
when converting to GridLayout, and adding them when converting from a
GridLayout.

When changing layouts, any layout params that are supported on the old
parent layout, but not the new layout, are removed. However, this was
not done for the non-layout attributes on the parent layout.  This
changeset also fixes that bug, so for example after converting a
horizontal LinearLayout to a RelativeLayout, the "orientation"
attribute is removed.

Change-Id: Id25b80b567578bd5882c527bbe23115577ff7d61