Support for palette configurations in Change Layout Refactoring
Some items in the palette are just variations of other items with some
different initial attributes. The Wrap In refactoring already supports
these, but the Change Layout did not - so for example converting to a
Vertical Linear Layout would not set the orientation="vertical"
attribute. This changeset adds support for these in the change layout
refactoring as well.

Change-Id: I9b44a57ebf5cddba7ae381b48b666e53f7f41ab0