Fix code completion filtering on Eclipse 3.7
The builtin Eclipse XML code completion provides some bad results. For
example, it will always offer the current value of the attribute,
regardless of whether it is valid. When you are typing in a value and
invoke code completion, this means it will offer you the prefix you've
already typed.

To get rid of this, we add our own "filtering" code completion which
hides these results. We can't just completely replace the XML code
completion providers since we want some results, such as template
completion suggestions. Instead we look for the specific XML code
completion provider and filter it out.

In Eclipse 3.7, the class we're filtering out is no longer used and
there's a new code completion provider. We need to filter that one out
too, and that is what this changeset does. It cannot reference the new
class directly since it does not exist in Eclipse 3.5, so we match by
class name instead.

Change-Id: Id8fae237a43a6b0dad34a148ee4f99729f1cbd63