Load standalone system images in the SdkManager.
This revamps the way system images are handled
in the internal SdkManager class.

Before, a given IAndroidTarget could provide a list
of ABI strings it new about, discovered by parsing
the SDK/platform/images/xyz or SDK/addon/images/xyz
folder.

This introduces the notion of System Image to an
IAndroidTarget. A system image combines an ABI with
a location strategy (legacy images folder, images
sub-folder or standalone sdk/system-images folder)
and an actual location path.

Change-Id: If5b748aa9aef6788bc3c814818381c7918b40bca