Support property sheet customizers on Linux (issue #18348)
On Linux, clicking on the "..." button to open the customizer dialogs
for references or flags would simply not work. This seems to be due to
how focus is handled on Linux, and in particular, the selection
dispatch is never called on the button because the focus loss event
handling disables it.

This changeset works around this by removing deactivation from focus
loss events, on Linux only.

Change-Id: Idb256e490279c3f0234bbb09342e89802b09fdc3