Create logcat table with SWT.VIRTUAL.
Without SWT.VIRTUAL, the UI responsiveness goes down significantly
when large numbers of logcat messages are received.

Unfortunately, enabling it causes SWT to make all rows to be of
the same height, thereby clipping any long messages. In such a case,
the complete message can be viewed in the tooltip.

Change-Id: I78d6701ccac6ad94675e1ed707700eabfe34bce8