Improve validation error messages when SDKs are loading
If you bring up the "Create New Sample Project" wizard in an Eclipse
that has no open editors (so nothing has triggered an SDK load yet),
you will get an error message that is a bit misleading (that the
selected SDK has no samples), and even when the SDK finishes loading
the validation isn't updated.

This changeset improves this situation such that there is a message
during loading stating to wait, which gets cleared once it's done.

Change-Id: I5d0798eff3b4fea9c8835c2f88c5bc430110aac0