Fix SDK target selector to avoid horizontal scrollbar
This changeset works around a Mac SWT issue where a checkbox table
will not take into account the size of the checkbox column, so you end
up with a horizontal scrollbar. To work around this, we compute the
size of checkboxes and subtract it from the available width when
distributing column sizes.

Change-Id: I9c03b987f57bc992487b67d24215134ae0d5b0b8