AVD Launcher.exe for the Windows SDK.
In Tools R14, the Windows SDK will have 2 icons in the root
of the SDK: one to launch the AVD window, and a different one
to launche the SDK Manager window.

I'll need to change the icon in a later CL. I don't have
any ICO editor at hand right now.

Change-Id: Ic651e3c3d71c2eac5fe851fba8ea2782eca5e594