Support grouping of allocations by library.
Currently, all allocations are displayed as a list sorted by
the size of the allocation.

This patch set enables the toolbar checkbox to group the list of
allocations by library. When this is checked, the display refreshes
to show a tree with the root nodes being libraries, and their
children being the list of allocations made from that library.

Change-Id: Ibb61242db8c6e58d300f767f0b2b11b6b7a9530d