Cherrypick f37031 from master. do not merge.
Clarify SDK loading message

When SDKs are still loading there is a "please wait" message in the
status window. It would say "The project target (Android 2.3) ..."
when loading rendering targets, which was a bit confusing. The message
will now say "project target" when loading the target that is the
current project target, and otherwise the "rendering target".

Also change the refresh message to just say "The layout will refresh"
instead of mentioning it by name since when the layout name is
lowercase (as it usually is) the sentence begins with a lowercase
letter which looks a bit wrong (and besides there is no ambiguity
about which layout is going to be reloaded).

Change-Id: I8d6a456c2c8ba5fe539d07eef1f5a017c9b8454a