Also process and cache the png files in libraries.
The png in the libraries were not processed by the crunch step.
Because aapt never processes png files anymore (relying on the crunch
step to do it), the png files in libraries were never processed.

While this is less a problem for standard png files, this completely
breaks 9-patches that must be processed to actually behaves as 9-patch.

Change-Id: I0a1fe14ea34f6d36a368b456494410945afc3c44