Cherrypick 25f755 from master. do not merge.
Display tooltips only for long messages.

Currently, the tooltips are displayed for all logcat messages.
This patch makes them show up only if the logcat messages are longer
than the width of the column in which they are displayed, i.e.,
tooltips will be displayed only for those messages that could be clipped.

Change-Id: Iade7f6628a93a9344b337e45b957ca53e6fa4595